requirement.txt
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
Pillow==9.1.1
djangorestframework-simplejwt==5.1.0
